brique-61/requirements.txt
# Brique 61 - Python ML Dependencies

# Core ML
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Cloud Storage
boto3==1.34.10
botocore==1.34.10

# API Server (FastAPI)
fastapi==0.105.0
uvicorn[standard]==0.25.0
pydantic==2.5.2

# ML Explainability
shap==0.44.0

# Utilities
python-dotenv==1.0.0
joblib==1.3.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
